// dashboard/app/dashboard/users/[id]/type.ts
export interface User {
  id: string;
  name: string;
  email: string;
  emeralds: number;
  streakShields: number;
  xp: number;
  avatar: string;
  avatarHash: string;
  voiceMessages: number;
  tier: string;
  timezone: string;
  timeZoneOffset: number;
  createdAt: string;
  updatedAt: string;
}